Afegeix el requeriment de la contrasenya a l'hora d'esborrar un compte
A l'hora d'esborrar un compte, ara demanara l'usuari i la contrasenya.
S'han afegit els tests necessaris per a provar aquest codi.

+++ src/main/java/PetSitters/service/PetSittersService.java
@@ -1,11 +1,10 @@
 package PetSitters.service;
 
+import PetSitters.entity.UserPetSitters;
 import PetSitters.exception.ExceptionInvalidAccount;
 import PetSitters.schemas.DeleteAccountSchema;
-import PetSitters.entity.User;
 import PetSitters.repository.UserRepository;
 import PetSitters.schemas.LoginSchema;
-import PetSitters.schemas.LogoutSchema;
 import PetSitters.schemas.RegisterSchema;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
@@ -20,31 +19,34 @@ public class PetSittersService {
     @Autowired
     UserRepository UserRep;
 
-    public List<User> login(LoginSchema login) {
-        User test = new User(login.getUser(),login.getPassword());
+    public List<UserPetSitters> login(LoginSchema login) {
+        UserPetSitters test = new UserPetSitters(login.getUser(),login.getPassword());
         test.setId("1");
         UserRep.save(test);
         return UserRep.findByFirstName("why");
-
-    }
-
-    public void logout(LogoutSchema logout) {
     }
 
     public void register(RegisterSchema register) throws ParseException {
         register.validate();
-        User newUser = new User(register);
+        UserPetSitters newUser = new UserPetSitters(register);
         UserRep.save(newUser);
     }
 
     public void deleteAccount(DeleteAccountSchema account) throws ExceptionInvalidAccount {
         account.validate();
         String username = account.getUsername();
+        String password = account.getPassword();
 
-        User u = UserRep.findByUsername(username);
+        UserPetSitters u = UserRep.findByUsername(username);
+
         if (u == null) {
             throw new ExceptionInvalidAccount("The account with the username '" + username + "' does not exist");
         }
+
+        if (!u.isTheSamePassword(password)) {
+            throw new ExceptionInvalidAccount("The username or password provided are incorrect");
+        }
+
         UserRep.deleteByUsername(username);
     }
 }
